Move MessageReference to a different package
