+++ app/core/src/main/java/com/fsck/k9/activity/ChooseFolder.java
@@ -23,6 +23,7 @@ import com.fsck.k9.Account;
 import com.fsck.k9.Account.FolderMode;
 import com.fsck.k9.K9;
 import com.fsck.k9.Preferences;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.core.R;
 import com.fsck.k9.controller.MessagingController;
 import com.fsck.k9.controller.MessagingListener;
