+++ app/core/src/main/java/com/fsck/k9/activity/MessageLoaderHelper.java
@@ -17,6 +17,7 @@ import android.support.v4.content.Loader;
 import com.fsck.k9.Account;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.autocrypt.AutocryptOperations;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.controller.MessagingController;
 import com.fsck.k9.controller.MessagingListener;
 import com.fsck.k9.controller.SimpleMessagingListener;
