+++ app/core/src/main/java/com/fsck/k9/activity/NotificationDeleteConfirmation.java
@@ -15,12 +15,13 @@ import android.support.annotation.NonNull;
 import com.fsck.k9.Account;
 import com.fsck.k9.K9;
 import com.fsck.k9.Preferences;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.core.R;
 import com.fsck.k9.controller.MessagingController;
 import com.fsck.k9.notification.NotificationActionService;
 
-import static com.fsck.k9.activity.MessageReferenceHelper.toMessageReferenceList;
-import static com.fsck.k9.activity.MessageReferenceHelper.toMessageReferenceStringList;
+import static com.fsck.k9.controller.MessageReferenceHelper.toMessageReferenceList;
+import static com.fsck.k9.controller.MessageReferenceHelper.toMessageReferenceStringList;
 
 
 public class NotificationDeleteConfirmation extends Activity {
