+++ app/core/src/main/java/com/fsck/k9/activity/compose/MessageActions.java
@@ -7,7 +7,7 @@ import android.os.Parcelable;
 import com.fsck.k9.Account;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.activity.MessageCompose;
-import com.fsck.k9.activity.MessageReference;
+import com.fsck.k9.controller.MessageReference;
 
 public class MessageActions {
     /**
