+++ app/core/src/main/java/com/fsck/k9/controller/MessageReference.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.activity;
+package com.fsck.k9.controller;
 
 
 import java.util.StringTokenizer;
