+++ app/core/src/main/java/com/fsck/k9/mailstore/LocalFolder.java
@@ -28,7 +28,7 @@ import android.support.annotation.NonNull;
 
 import com.fsck.k9.Account;
 import com.fsck.k9.K9;
-import com.fsck.k9.activity.MessageReference;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.activity.Search;
 import com.fsck.k9.backend.api.MessageRemovalListener;
 import com.fsck.k9.helper.FileHelper;
