+++ app/core/src/main/java/com/fsck/k9/notification/NewMailNotifications.java
@@ -8,7 +8,7 @@ import android.util.SparseArray;
 import com.fsck.k9.Account;
 import com.fsck.k9.K9;
 import com.fsck.k9.K9.NotificationHideSubject;
-import com.fsck.k9.activity.MessageReference;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.mailstore.LocalMessage;
