+++ app/core/src/main/java/com/fsck/k9/notification/NotificationContent.java
@@ -1,7 +1,7 @@
 package com.fsck.k9.notification;
 
 
-import com.fsck.k9.activity.MessageReference;
+import com.fsck.k9.controller.MessageReference;
 
 
 class NotificationContent {
