+++ app/core/src/main/java/com/fsck/k9/notification/NotificationContentCreator.java
@@ -8,7 +8,7 @@ import android.text.TextUtils;
 import com.fsck.k9.Account;
 import com.fsck.k9.K9;
 import com.fsck.k9.core.R;
-import com.fsck.k9.activity.MessageReference;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.helper.Contacts;
 import com.fsck.k9.helper.MessageHelper;
 import com.fsck.k9.mail.Address;
