+++ app/core/src/main/java/com/fsck/k9/notification/WearNotifications.java
@@ -12,7 +12,7 @@ import android.support.v4.app.NotificationCompat.WearableExtender;
 import com.fsck.k9.Account;
 import com.fsck.k9.K9;
 import com.fsck.k9.core.R;
-import com.fsck.k9.activity.MessageReference;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.controller.MessagingController;
