+++ app/core/src/main/java/com/fsck/k9/provider/RawMessageProvider.java
@@ -19,7 +19,7 @@ import android.support.annotation.Nullable;
 
 import com.fsck.k9.Account;
 import com.fsck.k9.Preferences;
-import com.fsck.k9.activity.MessageReference;
+import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.mail.FetchProfile;
 import com.fsck.k9.mail.Folder;
 import com.fsck.k9.mail.MessagingException;
