+++ app/core/src/test/java/com/fsck/k9/controller/MessageReferenceTest.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.activity;
+package com.fsck.k9.controller;
 
 
 import com.fsck.k9.mail.Flag;
